print usage
git-svn-id: https://plowshare.googlecode.com/svn/trunk@121 51fab9ec-ea3e-11dd-80ed-bda8ba2650cd

+++ googlecode-upload.sh
@@ -11,6 +11,7 @@ DIRECTORY=$PROJECT-$VERSION
 IFS=":" read USERNAME PASSWORD < "$AUTHFILE"
 FILE=$DIRECTORY.tgz
 
+test $# -ne 0 || { echo "Usage: $0 LOG"; exit 1; }
 rm -rf $DIRECTORY
 svn export $TRUNK $DIRECTORY --username $USERNAME --force
 tar -zcf $FILE $DIRECTORY
